Add core library and refactor namespaces
