+++ Case.sln
@@ -7,6 +7,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PluckList", "Case\PluckList
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Web", "Web\Web.csproj", "{6E602E6F-E8BE-4163-8BD7-B830FB63CB10}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Core\Core.csproj", "{D555E8C4-37AE-43C7-B935-79FE002FA103}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{6E602E6F-E8BE-4163-8BD7-B830FB63CB10}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6E602E6F-E8BE-4163-8BD7-B830FB63CB10}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6E602E6F-E8BE-4163-8BD7-B830FB63CB10}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D555E8C4-37AE-43C7-B935-79FE002FA103}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D555E8C4-37AE-43C7-B935-79FE002FA103}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D555E8C4-37AE-43C7-B935-79FE002FA103}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D555E8C4-37AE-43C7-B935-79FE002FA103}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
