+++ Case/src/Printer/ConsolePrinter.cs
@@ -1,0 +1,7 @@
+﻿namespace PluckList.Printer
+{
+    public abstract class ConsolePrinter : IPrint
+    {
+        public abstract void Print(string text = "default");
+    }
+}
