+++ Case/src/Printer/IPrint.cs
@@ -1,0 +1,7 @@
+﻿namespace PluckList.Printer
+{
+    public interface IPrint
+    {
+        void Print(string text);
+    }
+}
