+++ Case/src/Printer/OptionPrinter.cs
@@ -1,10 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace PluckList.src.Printer
+namespace PluckList.Printer
 {
     public class OptionPrinter : ConsolePrinter
     {
