--- Case/src/printer/IPrint.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PluckList.src.Printer
-{
-    public interface IPrint
-    {
-        void Print(string text);
-    }
-}
