+++ Web/Views/Home/Items.cshtml
@@ -1,5 +1,4 @@
-﻿@using PluckList
-
+﻿@using Core.Models
 <div>
     <h2>Items</h2>
     <hr/>
